Pulling out fixtures to conftest.py
